+++ modules/codec/opus.c
@@ -349,9 +349,9 @@ static int ProcessHeaders( decoder_t *p_dec )
     decoder_sys_t *p_sys = p_dec->p_sys;
     ogg_packet oggpacket;
 
-    unsigned pi_size[XIPH_MAX_HEADER_COUNT];
+    size_t pi_size[XIPH_MAX_HEADER_COUNT];
     const void *pp_data[XIPH_MAX_HEADER_COUNT];
-    unsigned i_count;
+    size_t i_count;
 
     size_t i_extra = p_dec->fmt_in->i_extra;
     const uint8_t *p_extra = p_dec->fmt_in->p_extra;
