+++ modules/demux/avformat/mux.c
@@ -203,11 +203,11 @@ static int AddStream( sout_mux_t *p_mux, sout_input_t *p_input )
         return VLC_EGENERIC;
     }
 
-    unsigned opus_size[XIPH_MAX_HEADER_COUNT];
+    size_t opus_size[XIPH_MAX_HEADER_COUNT];
     const void *opus_packet[XIPH_MAX_HEADER_COUNT];
     if( fmt->i_codec == VLC_CODEC_OPUS )
     {
-        unsigned count;
+        size_t count;
         /* Only transmits the first packet (OpusHead) */
         if( xiph_SplitHeaders(opus_size, opus_packet, &count, fmt->i_extra, fmt->p_extra ) ) {
             count = 0;
